fix: move component to its rightful place
